Tests avec EF (Select et Insert), configuration Map

+++ LaboratoireWebEntityFramework/Models/Interface/ITable.cs
@@ -8,7 +8,7 @@ namespace LaboratoireWebEntityFramework.Models.Interface
 {
     public interface ITable
     {
-        UInt64 Id { get; set; }
+        Int32 Id { get; set; }
         DateTime? DateCreation { get; set; }
         bool Actif { get; set; }
     }
